fix: sec/ns constant is now i64 by default

+++ sve2/utils/threads.h
@@ -8,7 +8,7 @@
 // thread timer, used to identify deadlines
 void init_threads_timer();
 i64 threads_timer_now();
-#define SVE2_NS_PER_SEC 1000000000
+#define SVE2_NS_PER_SEC ((i64)1000000000)
 
 #define sve2_thrd_create(...) nassert(thrd_create(__VA_ARGS__) == thrd_success)
 #define sve2_mtx_init(m, type) nassert(mtx_init(m, type) == thrd_success)
